more cleanup
:heavy_check_mark: handling empty cases
:heavy_check_mark: moving all function to helper script
